+++ okitweb/static/okit/properties/js/resources/block_storage_volume.js
@@ -37,12 +37,13 @@ class BlockStorageVolumeProperties extends OkitResourceProperties {
         this.size_in_gbs = size_in_gb.input
         this.append(this.size_tbody, size_in_gb.row)
         // Performance
-        const vpus_per_gb = this.createInput('number', 'Volume Performance', `${self.id}_vpus_per_gb`, '', (d, i, n) => {self.resource.vpus_per_gb = n[i].value; self.vpus_per_gb_range.property('value', self.resource.vpus_per_gb); this.setPerformanceTitle()}, {min: 0, max: 120, step: 10})
+        const vpus_per_gb_data = this.resource.isOCI() ? {min: 0, max: 120, step: 10} : {min: 0, max: 20, step: 20}
+        const vpus_per_gb = this.createInput('number', 'Volume Performance', `${self.id}_vpus_per_gb`, '', (d, i, n) => {self.resource.vpus_per_gb = n[i].value; self.vpus_per_gb_range.property('value', self.resource.vpus_per_gb); this.setPerformanceTitle()}, vpus_per_gb_data)
         this.vpus_per_gb = vpus_per_gb.input
         this.vpus_per_gb_label = vpus_per_gb.title
         this.vpus_per_gb_row = vpus_per_gb.row
         this.append(this.size_tbody, vpus_per_gb.row)
-        const vpus_per_gb_range = this.createInput('range', '', `${self.id}_vpus_per_gb_range`, '', (d, i, n) => {self.resource.vpus_per_gb = n[i].value; self.vpus_per_gb.property('value', self.resource.vpus_per_gb); this.setPerformanceTitle()}, {min: 0, max: 120, step: 10})
+        const vpus_per_gb_range = this.createInput('range', '', `${self.id}_vpus_per_gb_range`, '', (d, i, n) => {self.resource.vpus_per_gb = n[i].value; self.vpus_per_gb.property('value', self.resource.vpus_per_gb); this.setPerformanceTitle()}, vpus_per_gb_data)
         this.vpus_per_gb_range = vpus_per_gb_range.input
         this.vpus_per_gb_range_row = vpus_per_gb_range.row
         this.append(this.size_tbody, vpus_per_gb_range.row)
@@ -72,10 +73,10 @@ class BlockStorageVolumeProperties extends OkitResourceProperties {
         this.setPerformanceTitle()
         this.loadReferenceSelect(this.backup_policy, 'getVolumeBackupPolicies', true, undefined, {'Oracle-Defined Policies': this.oci_defined_filter, 'User-Defined Policies': this.user_defined_filter}, 'No Backup Policy Selected')
         this.backup_policy.property('value', this.resource.backup_policy)
-        if (this.resource.getOkitJson().metadata.platform === 'pca') {
-            this.vpus_per_gb_row.classed('collapsed', true)
-            this.vpus_per_gb_range_row.classed('collapsed', true)
-        }
+        // if (this.resource.getOkitJson().metadata.platform === 'pca') {
+        //     this.vpus_per_gb_row.classed('collapsed', true)
+        //     this.vpus_per_gb_range_row.classed('collapsed', true)
+        // }
     }
 
     // Set Performance Label